extract SubmissionDetailContent
